split Stop and Departure tests into extra files
