+++ Example/Tests/StopTests.swift
@@ -1,0 +1,27 @@
+//
+//  StopTests.swift
+//  DVB
+//
+//  Created by Kilian Költzsch on 28/09/2016.
+//  Copyright © 2016 CocoaPods. All rights reserved.
+//
+
+import Quick
+import Nimble
+import DVB
+
+class StopTests: QuickSpec {
+    override func spec() {
+        describe("Stop") {
+            let stop = Stop(id: 0, name: "Helmholtzstraße", region: "Dresden", searchString: "", tarifZones: "", longitude: 1.0, latitude: 1.0, priority: 1)
+
+            it("should have the correct name") {
+                expect(stop.name) == "Helmholtzstraße"
+            }
+
+            it("should have the correct location") {
+                expect(stop.region) == "Dresden"
+            }
+        }
+    }
+}
